Minor refactor of serv/repo
Still unsure.  Feel like a frankenstein between
two patterns.
