+++ Slimer.Infrastructure/Repositories/Api/GitHubRepository.cs
@@ -19,7 +19,7 @@ namespace Slimer.Infrastructure.Repositories.Api
             _secretsService = secretsService ?? throw new ArgumentNullException(nameof(secretsService));
         }
 
-        public async Task<GitHubResponse> CreateIssueAsync(GitHubRequest request)
+        public async Task<GitHubResponse> PostIssueAsync(GitHubRequest request)
         {
             var httpRequest = CreateRequest(CreateContent(request), "https://api.github.com/repos/SpookyWitchcraft/BabaYaga/issues");
