+++ Slimer.Infrastructure/Repositories/Api/Interfaces/IChatGptRepository.cs
@@ -1,0 +1,7 @@
+﻿namespace Slimer.Infrastructure.Repositories.Api.Interfaces
+{
+    public interface IChatGptRepository
+    {
+        Task<IEnumerable<string>> GetAnswerAsync(string question);
+    }
+}
