+++ Slimer.Infrastructure/Repositories/Api/Interfaces/IMarvelRepository.cs
@@ -1,0 +1,9 @@
+﻿using Slimer.Domain.Contracts.Marvel;
+
+namespace Slimer.Infrastructure.Repositories.Api.Interfaces
+{
+    public interface IMarvelRepository
+    {
+        Task<MarvelDataResults> GetCharacterDetailsAsync(string name);
+    }
+}
