+++ Slimer/Services/Interfaces/IMarvelService.cs
@@ -4,6 +4,6 @@ namespace Slimer.Services.Interfaces
 {
     public interface IMarvelService
     {
-        Task<MarvelDataResults> GetCharacterDetailsAsync(string name);
+        Task<MarvelDataResults> SearchForCharacterDetailsAsync(string name);
     }
 }
